backend/backend-requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
aiohttp==3.9.1
fastapi-cache2==0.2.1
redis==5.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1
httpx==0.25.2
asyncio==3.4.3
typing_extensions==4.8.0
starlette==0.27.0
ruff==0.1.6
black==23.11.0
